feedback api

+++ FeedBackApp/src/screens/setting/index.js
@@ -1,22 +1,51 @@
-import {View, Text, SafeAreaView, Image, TouchableOpacity} from 'react-native';
-import React from 'react';
+import {
+  View,
+  Text,
+  SafeAreaView,
+  Image,
+  TouchableOpacity,
+  Platform,
+} from 'react-native';
+import React, {useState, useEffect} from 'react';
 import styles from './styles';
 import {CustomButtonComponent, CustomHeader} from '../../components';
 import {ImagePath} from '../../assets/images';
 import {Route} from '../../navigation/route';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import {COLOR} from '../../utils/color/color';
 const Setting = props => {
+  const [user, setUser] = useState('');
   const onSignOut = async () => {
-    await AsyncStorage.clear()
-      .then(() => {
-        console.log('async items cleared');
-        props.navigation.navigate('Login');
-      })
-      .catch(error => {
-        console.log('Error while Sign  Out', error);
-      });
+    try {
+      const asyncStorageKeys = await AsyncStorage.getAllKeys();
+      if (asyncStorageKeys.length > 0) {
+        if (Platform.OS === 'android') {
+          await AsyncStorage.clear();
+          console.log('async items cleared android');
+          props.navigation.navigate('Login');
+        }
+        if (Platform.OS === 'ios') {
+          await AsyncStorage.multiRemove(asyncStorageKeys);
+          console.log('async items cleared from ios');
+          props.navigation.navigate('Login');
+        }
+      }
+    } catch (error) {
+      console.log('Error while sign out', error);
+    }
   };
+  useEffect(() => {
+    getUser();
+  }, []);
+
+  async function getUser(params) {
+    const AsyncUser = await AsyncStorage.getItem('user');
+    // setUser();
+    let parsedUser = JSON.parse(AsyncUser);
+    setUser(parsedUser);
+    console.log('user from setting screen ', user);
+  }
   return (
     <CustomHeader>
       <View style={styles.mainView}>
@@ -51,7 +80,9 @@ const Setting = props => {
             source={ImagePath.ELLIPSE}
           />
           <View style={{width: '60%'}}>
-            <Text style={{fontSize: 18}}>Sam Perez</Text>
+            <Text style={{fontSize: 18, color: COLOR.TEXTBLACK}}>
+              {user?.businessName}
+            </Text>
             <View style={{flexDirection: 'row'}}>
               <Image source={ImagePath.LOCATION} />
               <Text style={{color: 'grey'}}>Jakarta, Indonesia</Text>
